refactor: split up component into focused parts
